Isolating startTxSpendingFromMultisig function
